Fix: Moved the class ItemWindow out of __init__.py. That required me to move out a bunch of functions too, including functions that are now in badges_functions.py, some functions in pokemon_functions.py, dome functions in business.py, some functions in utils.py
