+++ src/Ankimon/functions/pokemon_functions.py
@@ -1,7 +1,13 @@
-from ..resources import pokedex_path, next_lvl_file_path
 import csv
 import json
 import random
+import uuid
+from datetime import datetime
+
+from .pokedex_functions import search_pokeapi_db_by_id, search_pokedex, search_pokedex_by_id
+from .battle_functions import calculate_hp
+
+from ..resources import pokedex_path, next_lvl_file_path, mypokemon_path, learnset_path
 
 def pick_random_gender(pokemon_name):
     """
@@ -102,19 +108,12 @@ def find_experience_for_level(group_growth_rate, level, remove_levelcap=True):
         elif group_growth_rate == "slow":
             experience = ((5 * ((level+1) ** 3)) // 4) - ((5 * (level ** 3)) // 4)
         return experience
-    
-
-import random
 
 def shiny_chance():
     # Shiny Pokémon probability (1 in 4096 chance)
     SHINY_PROBABILITY = 4096
     shiny = random.randint(1, SHINY_PROBABILITY) == 1
     return shiny
-
-import uuid
-import json
-from datetime import datetime
 
 def create_caught_pokemon(enemy_pokemon, nickname):
     enemy_pokemon.stats["xp"] = 0
@@ -151,3 +150,142 @@ def create_caught_pokemon(enemy_pokemon, nickname):
         "special-form": None,
     }
     return caught_pokemon
+
+def get_random_moves_for_pokemon(pokemon_name, level):
+        """
+        Get up to 4 random moves learned by a Pokémon at a specific level and lower, along with the highest level,
+        excluding moves that can be learned at a higher level.
+
+        Args:
+            json_file_name (str): The name of the JSON file containing Pokémon learnset data.
+            pokemon_name (str): The name of the Pokémon.
+            level (int): The level at which to check for moves.
+
+        Returns:
+            list: A list of up to 4 random moves and their highest levels.
+        """
+        # Load the JSON file
+        with open(learnset_path, "r", encoding="utf-8") as file:
+            learnsets = json.load(file)
+
+        # Normalize the Pokémon name to lowercase for consistency
+        pokemon_name = pokemon_name.lower()
+
+        # Retrieve the learnset for the specified Pokémon
+        pokemon_learnset = learnsets.get(pokemon_name, {})
+
+        # Create a dictionary to store moves and their corresponding highest levels
+        moves_at_level_and_lower = {}
+
+        # Loop through the learnset dictionary
+        for move, levels in pokemon_learnset.get('learnset', {}).items():
+            highest_level = float('-inf')  # Initialize with negative infinity
+            eligible_moves = []  # Store moves eligible for inclusion
+
+            for move_level in levels:
+                # Check if the move_level string contains 'L'
+                if 'L' in move_level:
+                    # Extract the level from the move_level string
+                    move_level_int = int(move_level.split('L')[1])
+
+                    # Check if the move can be learned at the specified level or lower
+                    if move_level_int <= level:
+                        # Update the highest_level if a higher level is found
+                        highest_level = max(highest_level, move_level_int)
+                        eligible_moves.append(move)
+
+            # Check if the eligible moves can be learned at a higher level
+            if highest_level != float('-inf'):
+                can_learn_at_higher_level = any(
+                    int(move_level.split('L')[1]) > highest_level
+                    for move_level in levels
+                    if 'L' in move_level
+                )
+                if not can_learn_at_higher_level:
+                    moves_at_level_and_lower[move] = highest_level
+
+        attacks = []
+        if moves_at_level_and_lower:
+            # Convert the dictionary into a list of tuples for random selection
+            moves_and_levels_list = list(moves_at_level_and_lower.items())
+            random.shuffle(moves_and_levels_list)
+
+            # Pick up to 4 random moves and append them to the attacks list
+            for move, highest_level in moves_and_levels_list[:4]:
+                #attacks.append(f"{move} at level: {highest_level}")
+                attacks.append(f"{move}")
+
+        return attacks
+
+def save_fossil_pokemon(pokemon_id):
+    # Create a dictionary to store the Pokémon's data
+    # add all new values like hp as max_hp, evolution_data, description and growth rate
+    name = search_pokedex_by_id(pokemon_id)
+    id = pokemon_id
+    stats = search_pokedex(name, "baseStats")
+    abilities = search_pokedex(name, "abilities")
+    evos = search_pokedex(name, "evos")
+    gender = pick_random_gender(name.lower())
+    numeric_abilities = {k: v for k, v in abilities.items() if k.isdigit()}
+    # Check if there are numeric abilities
+    if numeric_abilities:
+        # Convert the filtered abilities dictionary values to a list
+        abilities_list = list(numeric_abilities.values())
+        # Select a random ability from the list
+        ability = random.choice(abilities_list)
+    else:
+        # Set to "No Ability" if there are no numeric abilities
+        ability = "No Ability"
+    type = search_pokedex(name, "types")
+    name = search_pokedex(name, "name")
+    generation_file = "pokeapi_db.json"
+    growth_rate = search_pokeapi_db_by_id(id, "growth_rate")
+    base_experience = search_pokeapi_db_by_id(id, "base_experience")
+    description= search_pokeapi_db_by_id(id, "description")
+    level = 5
+    attacks = get_random_moves_for_pokemon(name, level)
+    stats["xp"] = 0
+    ev = {
+        "hp": 0,
+        "atk": 0,
+        "def": 0,
+        "spa": 0,
+        "spd": 0,
+        "spe": 0
+    }
+    iv = {
+        "hp": random.randint(1, 32),
+        "atk": random.randint(1, 32),
+        "def": random.randint(1, 32),
+        "spa": random.randint(1, 32),
+        "spd": random.randint(1, 32),
+        "spe": random.randint(1, 32)
+    }
+    caught_pokemon = {
+        "name": name,
+        "gender": gender,
+        "level": level,
+        "id": id,
+        "ability": ability,
+        "type": type,
+        "stats": stats,
+        "ev": ev,
+        "iv": iv,
+        "attacks": attacks,
+        "base_experience": base_experience,
+        "current_hp": calculate_hp(int(stats["hp"]), level, ev, iv),
+        "growth_rate": growth_rate,
+        "evos": evos
+    }
+    # Load existing Pokémon data if it exists
+    if mypokemon_path.is_file():
+        with open(mypokemon_path, "r", encoding="utf-8") as json_file:
+            caught_pokemon_data = json.load(json_file)
+    else:
+        caught_pokemon_data = []
+
+    # Append the caught Pokémon's data to the list
+    caught_pokemon_data.append(caught_pokemon)
+    # Save the caught Pokémon's data to a JSON file
+    with open(str(mypokemon_path), "w") as json_file:
+        json.dump(caught_pokemon_data, json_file, indent=2)
